generate op chain modules with edr_op build script
