--- crates/edr_op/op_chains_importer/.gitignore
@@ -1,1 +1,0 @@
-target/
