--- crates/edr_op/op_chains_importer/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "op_chains_importer"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-git2 = "0.20.2"
-indexmap = "2.11.0"
-serde = "1.0.219"
-tempfile = "3.21.0"
-toml = {version = "0.9.5", features = ["preserve_order"]}
-
-
-[workspace]
